LayoutTests/fast/css/css-keyframe-style-parentRule.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
<style>
    @-webkit-keyframes anim {
        from {
            color: green;
        }
    }
</style>
</head>
<body>
<script>

description("This test verifies that a keyframe rule's 'parentRule' points back to the keyframe rule.");

var rules = document.styleSheets[1].cssRules;

shouldBe("rules[0][0].parentRule", "rules[0]");
shouldBe("rules[0][0].style.parentRule", "rules[0][0]");

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
