third_party/WebKit/LayoutTests/fast/css/css-keyframe-style-parentRule-expected.txt

This test verifies that a keyframe rule's 'parentRule' points back to the keyframe rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rules[0][0].parentRule is rules[0]
PASS rules[0][0].style.parentRule is rules[0][0]
PASS successfullyParsed is true

TEST COMPLETE
